Convert NFA representation to support tags (becoming tNFAs).
This forms the basis on which we can implement the algorithms outlined in Ville Laurikari's paper "NFAs with Tagged Transitions, their Conversion to Deterministic Automata and Application to Regular Expressions".
This allows building a DFA to match any regex with submatches, so we won't need to backtrack over smaller submatch DFAs anymore, but process everything in one go.