Fix #26 by taking the max of the inserted state numbers
When compiling an NFA for an alternation expression, if the left part
of the expression is an empty sequence (but NOT when it is a literal
epsilon!), the new state added for the entire expression would get
the state number of the empty sequence (which is the same as the
state number preceding the expression).  This would have the effect
of dropping the right part of the expression in the NFA.  But there
would still be references to them, causing strange errors.

Instead, we must take the maximum of the two, because we never know
which of the two expressions inserts new states.