tests: Add suite-name to test-begin.
The SRFI-64 defines the test-begin form of this shape:

    (test-begin suite-name [count])

Currently all test-begin present in the tests are missing the suite-name, thus
failing to conform.  Chicken does accept that, however for example Guile does
not.  Therefore this commit adds the suite-name everywhere.  In all instances it
is just set to the suffix of the test filename, so it does not really add much
information, however it will allow running the tests on SRFI-64 compliant
schemes.

* tests/test-cset.scm,
tests/test-irregex-from-gauche.scm,
tests/test-irregex-pcre.scm,
tests/test-irregex-scsh.scm,
tests/test-irregex-utf8.scm,
tests/test-irregex.scm: Add the suite-name parameter to test-begin form.