tests: Move tests into a new directory.
In preparation for splitting of the test files into a test logic and
dialect-specific setup parts, the tests are moved into a new directory, `tests'.

The shebangs are removed from all the files.  Only one was actually executable,
so I get the impression they were not invoked like that anyway.  Even
test-all.scm had a note to run it using `csi -s test-all.scm'.  It also
hard-coded the csi location to /usr/local/bin/sci, which is sub-optimal, however
probably the best one can do in chicken.

* manifest.scm: Add coreutils, for dirname.
* test-all.scm: Delete file.
* tests/all: Replacement for test-all.scm, written in shell.
* tests/test-cset.scm: Move from test-cset.scm.
* tests/test-irregex-from-gauche.scm: Move from test-irregex-gauche.scm.
* tests/test-irregex-pcre.scm: Move from test-irregex-pcre.scm.  Remove shebang.
* tests/test-irregex-scsh.scm: Move from test-irregex-scsh.scm.
* tests/test-irregex-utf8.scm: Move from test-irregex-utf8.scm.
* tests/test-irregex.scm: Move from test-irregex.scm.  Change mode to 0644.
Remove shebang.
* tests/README: New file.