tests: Add support for running tests on GNU Guile.
Straightforward changes, except it did require one change in the library itself,
and that is to turn off utf8 parsing support while running under guile.

Based on the information from Alex Shinn, it is useful only for Scheme
implementations that do not natively support unicode and instead consider
strings to be sequences of bytes, in this case in utf8 encoding.  For others, it
is harmful and can break processing of non-ascii text.

You can tell the type of your scheme interpreter by running the following code:

    (display (string-length "é"))
    (newline)
    (display (string-length "あ"))
    (newline)

Guile prints 1 and 1, while chicken 2 and 3.

In the end, the picked solution is to turn it off for everything that is not
chicken without full-unicode feature (chicken 6 will support it).

* .gitignore: Ignore the .log files created by Guile's test driver.
* irregex.scm (*allow-utf8-mode?*): Set to #f for everything except chicken < 6.
* manifest.scm: Add guile.
* tests/all (schemes): Register guile.
(interpret_guile): New function.  Run the guile with no auto-compilation due to
the startup cost.
* tests/guile-cset.scm,
tests/guile-irregex-from-gauche.scm,
tests/guile-irregex-pcre.scm,
tests/guile-irregex-scsh.scm,
tests/guile-irregex-utf8.scm,
tests/guile-irregex.scm,
tests/guile/test-support.scm: New files.